+++ src/components/UhiLogo.js
@@ -1,5 +1,6 @@
 import Logo from '../Assets/uhi-logo.png';
+import Image from 'next/image';
 function UhiLogo(){
-return <img src={Logo} alt="UHI Logo" class='h-96' />;
+return <Image src={Logo} alt="UHI Logo" class='h-96' />;
 }
 export default UhiLogo;
